+++ lib/api/input_api.dart
@@ -2,26 +2,25 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:intl/intl.dart';
-import 'crud_page_view.dart';
+import '../screens/input/input_screen.dart';
 
-class CrudPage extends StatefulWidget {
+class InputApi extends StatefulWidget {
   @override
-  _CrudPageState createState() => _CrudPageState();
+  _InputApiState createState() => _InputApiState();
 }
 
-class _CrudPageState extends State<CrudPage> {
+class _InputApiState extends State<InputApi> {
   FirebaseFirestore firestore = FirebaseFirestore.instance;
   FirebaseAuth auth = FirebaseAuth.instance;
   Stream<QuerySnapshot> currentStream;
 
   void initState() {
     super.initState();
-    currentStream =
-        firestore
-            .collection("user")
-            .doc(auth.currentUser.uid)
-            .collection(DateFormat('yyyy-MM').format(DateTime.now()))
-            .snapshots();
+    currentStream = firestore
+        .collection("user")
+        .doc(auth.currentUser.uid)
+        .collection(DateFormat('yyyy-MM').format(DateTime.now()))
+        .snapshots();
   }
 
   @override
@@ -32,7 +31,8 @@ class _CrudPageState extends State<CrudPage> {
         builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
           if (!snapshot.hasData) {
             return CircularProgressIndicator();
-          } return CrudPageView(snapshot.data.docs);
+          }
+          return InputScreen(snapshot.data.docs);
         },
       ),
     );
