+++ lib/screens/calculate_result/power_calculated_second_result_screen.dart
@@ -1,20 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'document_view.dart';
+import 'document_second_screen.dart';
 import 'package:flutter_swiper/flutter_swiper.dart';
+import '../../constants.dart';
 
-class MyPowerPageView extends StatelessWidget {
+class PowerCalculatedSecondResultScreen extends StatelessWidget {
   final num result;
   final String date;
   final List<DocumentSnapshot> documents;
-  MyPowerPageView(this.date, this.documents, this.result);
+  PowerCalculatedSecondResultScreen(this.date, this.documents, this.result);
 
   @override
   Widget build(BuildContext context) {
+    List<Widget> _contatos = new List<Widget>();
+    _contatos.add(DocumentSecondScreen('전자기기', '전력 소비량', '시간 순'));
+    _contatos += documents
+        .map((eachDocument) => DocumentSecondScreen(
+            eachDocument['device'],
+            eachDocument["calculate"].toInt().toString() + 'WH',
+            eachDocument['UsageTime'].toString()))
+        .toList();
+
     final _screenSize = MediaQuery.of(context).size;
     return new Scaffold(
-      resizeToAvoidBottomPadding: false,
+      resizeToAvoidBottomInset: false,
       extendBodyBehindAppBar: true,
       appBar: AppBar(
         backgroundColor: Colors.transparent,
@@ -32,18 +42,11 @@ class MyPowerPageView extends StatelessWidget {
           return new Container(
             width: double.infinity,
             height: double.infinity,
-            decoration: BoxDecoration(
-              gradient: LinearGradient(
-                colors: [Color(0xFFE7F3EB), Color(0xFFF8F5E1)],
-                begin: Alignment.topCenter,
-                end: Alignment.bottomCenter,
-              ),
-            ),
+            decoration: kRegularDecoration,
             child: new Swiper(
-                controller: new SwiperController(),
-                control: SwiperControl(
-                  color: Colors.black,
-                ),
+                // control: SwiperControl(
+                //   color: Colors.black,
+                // ),
                 itemCount: 200,
                 scale: 0.6,
                 viewportFraction: 0.9,
@@ -53,7 +56,7 @@ class MyPowerPageView extends StatelessWidget {
                     children: <Widget>[
                       new Center(
                         child: Text(
-                          '${_getLastMonth(date)}월보다 이번 달에\n ${(0 <= result) ? '+' + result.toInt().toString() : '-' + result.toInt().toString()} 사용했어요',
+                          '지난 달보다 ${_getMonth(index, date)}월에\n⚡ ${(0 <= result) ? '+' + result.toInt().toString() : result.toInt().toString()}WH ⚡\n사용했어요!',
                           textAlign: TextAlign.center,
                           style: TextStyle(
                             color: Colors.black,
@@ -62,21 +65,25 @@ class MyPowerPageView extends StatelessWidget {
                           ),
                         ),
                         heightFactor: _screenSize.height * 0.003,
-
                       ),
                       new Container(
                         decoration: BoxDecoration(
                           color: Colors.white,
                           borderRadius: BorderRadius.circular(25),
+                          boxShadow: [
+                            BoxShadow(
+                              color: Colors.grey,
+                              offset: Offset(2.0, 4.0),
+                              blurRadius: 5.0,
+                            ),
+                          ],
                         ),
                         height: _screenSize.height * 0.65,
                         width: _screenSize.width * 0.86,
                         child: new ListView(
-                          padding: EdgeInsets.all(30),
+                          padding: EdgeInsets.all(6),
                           scrollDirection: Axis.vertical,
-                          children: documents
-                              .map((eachDocument) => DocumentView(eachDocument))
-                              .toList(),
+                          children: _contatos,
                         ),
                       ),
                     ],
@@ -88,9 +95,19 @@ class MyPowerPageView extends StatelessWidget {
     );
   }
 
-  int _getLastMonth(String date) {
-    int month = int.parse(date.split("-")[1]);
-    if (month == 1) return 12;
-    return month - 1;
+  int _getMonth(int index, String date) {
+    if (index == 0) {
+      int month = int.parse(date.split("-")[1]);
+      if (month == 12) {}
+      return month;
+    } else if (index > 0) {
+      int month = int.parse(date.split("-")[1]) + index;
+      return month;
+    } else {
+      if (index == 0) {
+        int month = int.parse(date.split("-")[1]) - index;
+        return month;
+      }
+    }
   }
 }
